Extract abstract test class
