+++ core/trino-main/src/test/java/io/trino/memory/TestLeastWastedEffortTaskLowMemoryKiller.java
@@ -11,7 +11,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package io.trino.memory;
 
 import com.google.common.collect.ImmutableList;
@@ -44,66 +43,12 @@ import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.testng.Assert.assertEquals;
 
 public class TestLeastWastedEffortTaskLowMemoryKiller
+        extends AbstractTaskLowMemoryKillerTest
 {
-    private final LowMemoryKiller lowMemoryKiller = new LeastWastedEffortTaskLowMemoryKiller();
-
-    @Test
-    public void testMemoryPoolHasNoReservation()
+    @Override
+    protected LowMemoryKiller createLowMemoryKiller()
     {
-        int memoryPool = 12;
-        Map<String, Map<String, Long>> queries = ImmutableMap.of(
-                "q_1",
-                ImmutableMap.of("n1", 0L, "n2", 0L, "n3", 0L, "n4", 0L, "n5", 0L));
-
-        assertEquals(
-                lowMemoryKiller.chooseTargetToKill(
-                        toRunningQueryInfoList(queries),
-                        toNodeMemoryInfoList(memoryPool, queries)),
-                Optional.empty());
-    }
-
-    @Test
-    public void testMemoryPoolNotBlocked()
-    {
-        int memoryPool = 12;
-        Map<String, Map<String, Long>> queries = ImmutableMap.<String, Map<String, Long>>builder()
-                .put("q_1", ImmutableMap.of("n1", 0L, "n2", 6L, "n3", 0L, "n4", 0L, "n5", 0L))
-                .put("q_2", ImmutableMap.of("n1", 3L, "n2", 5L, "n3", 2L, "n4", 4L, "n5", 0L))
-                .buildOrThrow();
-        assertEquals(
-                lowMemoryKiller.chooseTargetToKill(
-                        toRunningQueryInfoList(queries),
-                        toNodeMemoryInfoList(memoryPool, queries)),
-                Optional.empty());
-    }
-
-    @Test
-    public void testWillNotKillTaskForQueryWithoutTaskRetriesEnabled()
-    {
-        int memoryPool = 5;
-        Map<String, Map<String, Long>> queries = ImmutableMap.<String, Map<String, Long>>builder()
-                .put("q_1", ImmutableMap.of("n1", 0L, "n2", 0L, "n3", 2L))
-                .put("q_2", ImmutableMap.of("n1", 3L, "n2", 5L, "n3", 2L))
-                .buildOrThrow();
-        Map<String, Map<String, Map<Integer, Long>>> tasks = ImmutableMap.<String, Map<String, Map<Integer, Long>>>builder()
-                .put("q_1", ImmutableMap.of(
-                        "n3", ImmutableMap.of(1, 5L)))
-                .put("q_2", ImmutableMap.of(
-                        "n1", ImmutableMap.of(
-                                1, 1L,
-                                2, 2L),
-                        "n2", ImmutableMap.of(
-                                3, 3L,
-                                4, 1L,
-                                5, 1L),
-                        "n3", ImmutableMap.of(6, 2L)))
-                .buildOrThrow();
-
-        assertEquals(
-                lowMemoryKiller.chooseTargetToKill(
-                        toRunningQueryInfoList(queries),
-                        toNodeMemoryInfoList(memoryPool, queries, tasks)),
-                Optional.empty());
+        return new LeastWastedEffortTaskLowMemoryKiller();
     }
 
     @Test
@@ -163,7 +108,7 @@ public class TestLeastWastedEffortTaskLowMemoryKiller
         }
 
         assertEquals(
-                lowMemoryKiller.chooseTargetToKill(
+                getLowMemoryKiller().chooseTargetToKill(
                         toRunningQueryInfoList(queries, ImmutableSet.of("q_1", "q_2"), taskInfos),
                         toNodeMemoryInfoList(memoryPool, queries, tasks)),
                 Optional.of(KillTarget.selectedTasks(
@@ -208,7 +153,7 @@ public class TestLeastWastedEffortTaskLowMemoryKiller
         // q2_3; n2; walltime 60s;  memory 2; ratio 0.033 (pick for n2)
 
         assertEquals(
-                lowMemoryKiller.chooseTargetToKill(
+                getLowMemoryKiller().chooseTargetToKill(
                         toRunningQueryInfoList(queries, ImmutableSet.of("q_1", "q_2"), taskInfos),
                         toNodeMemoryInfoList(memoryPool, queries, tasks)),
                 Optional.of(KillTarget.selectedTasks(
